move resizing to own file / helpers
